Moved AccountManager and login-related functionality (as well as Achievement and History - for now) to a new package called account
