+++ GroupProject/src/account/AccountManagerTest.java
@@ -1,4 +1,5 @@
-package QuizProject;
+//package QuizProject;
+package account;
 
 import static org.junit.Assert.*;
