+++ GroupProject/src/account/HistoryItem.java
@@ -1,4 +1,5 @@
-package QuizProject;
+//package QuizProject;
+package account;
 
 /**
  * Holds a single history record
