+++ GroupProject/src/account/LoginServlet.java
@@ -1,4 +1,5 @@
-package QuizProject;
+//package QuizProject;
+package account;
 
 import java.io.IOException;
